changed time from signed into to uint16_t

+++ timer.h
@@ -4,7 +4,7 @@
 #define HZ 100
 #define PRESCALE 64
 #define TIMER_TGT ( ( ( F_CPU / PRESCALE ) / HZ) - 1)
-volatile int jiffies;
+volatile uint16_t jiffies;
 void timer_init(void);
 
 #endif
